Rename `useFlowState` hook file to match the convention naming
