+++ packages/react/src/components/dynamic-flow/index.tsx
@@ -4,7 +4,7 @@ import { FormProvider } from "../../context/form-context";
 import { useCallback } from "react";
 import { Handle, LoginOptions } from "../../domain/types";
 import { CreateFlowOptions } from "../form/flow";
-import { useFlowState } from "../form/useFlowState";
+import { useFlowState } from "../form/use-flow-state";
 import { Authenticating } from "../form/authenticating";
 import { Success } from "../form/success";
 import { Error } from "../form/error";
